refactor:UserHasAttraction repository 수정 pk값 변경

+++ src/main/java/com/example/planergram/repository/UserHasAttractionRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface UserHasAttractionRepository extends JpaRepository<UserHasAttraction, Integer> {
+public interface UserHasAttractionRepository extends JpaRepository<UserHasAttraction, Long> {
 
     List<UserHasAttraction> findByUser(@Param("userId") int userId);
